fix(refactor): fix comments after review
