+++ src/javascripts/helpers/format.js
