+++ src/styles/components/popup.css
@@ -33,7 +33,7 @@
   padding: 24px 40px 64px 40px;
   flex-direction: column;
   align-items: flex-start;
-  gap: 24px;
+  gap: 10px;
 }
 
 .transaction-input-fields {
@@ -115,6 +115,7 @@
 
 .action-form {
   align-self: flex-end;
+  margin-top: 10px;
 }
 
 .btn-save,
@@ -146,4 +147,17 @@
   font-size: var(--font-size-sm);
   font-weight: 500;
   border: none;
+}
+.error-messages {
+  display: flex;
+  flex-direction: row;
+  gap: 165px;
+}
+
+.error-note {
+  margin-top: 10px;
+}
+
+.error {
+  visibility: hidden;
 }
