+++ src/styles/pages/home.css
@@ -76,8 +76,8 @@
   justify-content: space-between;
   border-bottom: 1px solid rgba(0, 0, 0, 0.12);
   height: 69px;
-  padding-right: 15.8px;
-  padding-left: 16px;
+  padding-right: 30px;
+  padding-left: 28px;
 }
 
 .gap {
